mcp_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.3
python-dotenv==1.0.1
pydantic==2.5.0
